+++ resources/js/Components/AppVersion/AppVersion.vue
@@ -11,7 +11,7 @@
             TV &#169;2009-{{new Date().getFullYear()}}
         </div>
         <div class="flex justify-center">
-            Beta Preview v0.7.3.5ec
+            Beta Preview v0.7.3.5ed
         </div>
         <div class="flex justify-center text-blue-600 hover:text-blue-400 underline">
             <Link @click="userStore.closeNavDropdown()" :href="route('changelog.show')">changelog</Link>